// src/app/Models/Product.ts
export class Product{
  
    id: number;
    name: string;
    category: string;
    brand: string;
    description: string;
    price: number;
    discount: number; // Include the discount property
    imageUrl: string;
    color: string[];
    itemLeft: number;
    is_In_inventory: boolean;
    slug: string;
    
  constructor(){

 }


}